Auto-reconnect and stop at interpunction
Prior to this, text was generated up to a certain number of words,
regardless of interpunction (a full stop was added at the end). The
current implementation is more natural, because it stops after the last
appropriate existing interpunction in generated text.

In addition, this commit solves issue #1 by automatically attempting to
reconnect to Twitter when {'hangup':True} is received in the
auto-replying Thread, or when the auto-tweeting Thread fails to post a
Tweet.