Allow for lists of prefixes, suffixes, and databases to be passed
The functions for auto-tweeting and for automatic replying now both
allow for passing lists for the `prefix` and `suffix` keyword. (Before,
only strings were allowed.) A random prefix or suffix will be selected
from the passed list. This completes the feature request in issue #3 .

In addition, the auto-tweeting now also allows for passing a list of
databases with the `database` keyword, which now mirrors the
functionality of the `database` keyword in the auto-replying function.