Bug fixes and new simple reply functionality
Added the option of giving pre-prepared replies using the
`set_simple_reply` method and `'simplereply'` database in
`twitter_autoreply_start`.

Fixed the bug in which None was passed to generate an automatic response
even after finding keywors.

Added a `_cpr` menthod that checks whether Threads are alive, and
revives them if required. This method is called in the `autoreplier`,
`autotweeter`, and the new `selfexaminer` Threads. If one or two fail,
the remaining Thread(s) will revive them. If all three Threads fail at
the same time, we're still screwed.